Some package restructuring, pulled the plug officially on result type.
