+++ package.json
@@ -1,12 +1,9 @@
 {
   "name": "ferp",
   "module": "src/ferp.js",
-  "main": "dist/ferp.js",
+  "main": "ferp.js",
   "description": "Build functional and pure applications in NodeJS and the Browser!",
   "version": "1.0.0-beta.2",
-  "directories": {
-    "example": "examples"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/mrozbarry/ferp.git"
